+++ test/test_tool_call.rb
@@ -36,12 +36,7 @@ class TestToolCall < KwardTestCase
     assert_equal({ timeoutSeconds: 7, nestedValue: [{ oldText: "old" }] }, Kward::ToolCall.camelize_args({ "timeout_seconds" => 7, "nested_value" => [{ "old_text" => "old" }] }))
   end
 
-  def test_classifies_write_lock_and_file_change_tools
-    assert Kward::ToolCall.write_lock_required?("edit_file")
-    assert Kward::ToolCall.write_lock_required?("write_file")
-    assert Kward::ToolCall.write_lock_required?("run_shell_command")
-    assert Kward::ToolCall.write_lock_required?("bash")
-    refute Kward::ToolCall.write_lock_required?("read_file")
+  def test_classifies_file_change_tools
 
     assert Kward::ToolCall.file_change_tool?("edit_file")
     assert Kward::ToolCall.file_change_tool?("write")